pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "mcp-rag"
version = "0.1.0"
description = "ragable ai agent with rag mcp server"
authors = [
    {name = "aorticweb", email = "aorticweb@aorticweb.com"}
]
license = {text = "MIT"}
requires-python = ">=3.12.9"
dependencies = []

[project.optional-dependencies]

embedder = [
    "pyyaml>=6.0.1",
    "sentence-transformers==5.0.0",
    "torch>=2.7.0",
]
server = [
    "fastmcp>=2.5.1",
    "faiss-cpu>=1.11.0",
    "pypdf2>=3.0.1",
    "sqlite-vec>=0.1.6",
    "sqlean-py>=3.49.1",
    "python-docx>=1.2.0",
    "python-pptx>=1.0.2",
    "beautifulsoup4>=4.13.4",
    "markdown>=3.8.2",
    "olefile>=0.47",
    "docx2txt>=0.9",
    "yt-dlp>=2025.6.30",
    "llvmlite>=0.44.0",
    "numba>=0.61.2",
    "openai-whisper>=20250625",
    "parakeet-mlx>=0.3.5",
]
dev = [
    "anyio>=4.2.0",
    "black==23.11.0",
    "flake8==6.0.0",
    "flake8-black>=0.3.6",
    "gevent>=23.9.1",
    "huggingface-hub[cli]>0.30.2",
    "isort==5.13.0",
    "mypy>=1.17.0",
    "pytest>=8.3.3",
    "pytest-asyncio>=0.23.2",
    "pytest-cov>=4.1.0",
    "shiv>=1.0.0",
]

[tool.mypy]
ignore_missing_imports = true

[tool.black]
line-length = 120


[tool.hatch.build.targets.wheel]
packages = ["common", "embedder", "server", "transcriber"]

